backend/businatch/apps/blog/urls.py
from rest_framework import routers
from django.urls import path, include
from .views import PostViewSet

router = routers.DefaultRouter()
router.register(r'posts', PostViewSet, basename='post')

urlpatterns = [
    path('', include(router.urls)),
]