backend/requirements.txt
Django
djangorestframework
python-dotenv
psycopg2-binary
django-cors-headers
stripe
mailchimp-marketing